Fix Journal.journals4mp(selected=True) returns only unique elements
This bug arises to the fact that both
    1) INTERNATIONAL JOURNAL OF ACAROLOGY
    2) PLATN ECOL DIVERS
in the JCR's journal title link to
    Journal(Plant Ecol Divers)
    Title:
               MedlineTA: Plant Ecol Divers
                   Title: {'International journal of acarology', 'Plant ecology & diversity'}
        ISOAbbreviations: {'Plant Ecol Divers', 'Int J Acarol'}
    ISSN:
                   Print: {ISSN('1755-0874'), ISSN('0164-7954')}
              Electronic:
                 Linking: 0164-7954
             NlmUniqueID: {'100955727', '101510994'}
                    URLs: https://ncbi.nlm.nih.gov/nlmcatalog/100955727/    https://ncbi.nlm.nih.gov/nlmcatalog/101510994/
                  Counts: 5

However, the 5 counts in this Journal are not assigned to any Communities - they only consist of 'for', 'and', 'so' kind of matches, which are all filtered before assignments.
Therefore, this bug has no effect on the results upto date.